Tests/LibWeb/Text/expected/css/CSSStyleDeclaration-custom-properties.txt

--foo= bar
--foo= baz important
--foo= blank
